increase number of iterations because of bug fixed
As we are now correctly make an independent copy of model weights for each parititon, we are now correctly averaging weights among bags. Previously in test, because there is only one executor, all bags are referencing to the exact same Model object thus rendering bagging void. With the cloneable fix, it now consistently takes more iterations to coverge on a single machine.