Add option to initialize model using quantiles
Using quantiles (e.g. 5th, 95th) instead of min/max to initialie spline can create better buckets for features that are unimodal but heavily skewed.
Broadcast model during feature flattening to minimize task deserialization time (this is crucial for partition sample to work).
Initialize feature as point when min and max is the same for Spline. This will make BIAS a point feature instead of a Spline feature.