Custom Serialization of Model
Enable model that is not implemented in aerosolve-core be serialized and deserialized via ModelFactory. This is achieved by storing Model class name as the model name in model header. This makes it possible to implemented AbstractModel interface in non-aerosolve code. For example, this can support wrapping spark.ml, PhotonML and other models from different libraries.